feat(legal): add GPL3 license
At this moment I cannot see what kinds of derivatives can be made that should not be open source given that this is mostly documentation.

Maybe there may be designs, and diagrams that are contributed which businesses want to use? 🤔 This repository wouldn't hold any source files for designs anyway, so the linked figma or miro files could contain their own license.